+++ src/LiveStreamingServerNet.Flv/Internal/Services/FlvClientFactory.cs
@@ -1,5 +1,6 @@
 ﻿using LiveStreamingServerNet.Flv.Internal.Contracts;
 using LiveStreamingServerNet.Flv.Internal.Services.Contracts;
+using LiveStreamingServerNet.Utilities.Contracts;
 using Microsoft.Extensions.Logging;
 
 namespace LiveStreamingServerNet.Flv.Internal.Services
