+++ src/LiveStreamingServerNet.Rtmp/Utilities/Containers/Contracts/IFlvWriter.cs
@@ -1,8 +1,11 @@
 ﻿using LiveStreamingServerNet.Utilities.Buffers.Contracts;
 
-namespace LiveStreamingServerNet.Flv.Internal.Contracts
+namespace LiveStreamingServerNet.Rtmp.Utilities.Containers.Contracts
 {
-    internal interface IFlvWriter : IAsyncDisposable
+    /// <summary>
+    /// Represents a FLV writer.
+    /// </summary>
+    public interface IFlvWriter : IAsyncDisposable
     {
         ValueTask WriteHeaderAsync(bool allowAudioTags, bool allowVideoTags, CancellationToken cancellationToken);
         ValueTask WriteTagAsync(FlvTagType tagType, uint timestamp, Action<IDataBuffer> payloadBuffer, CancellationToken cancellationToken);
