+++ src/LiveStreamingServerNet.Rtmp/Utilities/Containers/FlvTagHeader.cs
@@ -1,8 +1,21 @@
 ﻿using LiveStreamingServerNet.Utilities.Buffers.Contracts;
 
-namespace LiveStreamingServerNet.Flv.Internal
+namespace LiveStreamingServerNet.Rtmp.Utilities.Containers
 {
-    internal record struct FlvTagHeader(FlvTagType TagType, uint DataSize, uint Timestamp)
+    /// <summary>
+    /// Represents the type of an FLV tag.
+    /// </summary>
+    public enum FlvTagType : byte
+    {
+        Audio = 8,
+        Video = 9,
+        ScriptData = 18
+    }
+
+    /// <summary>
+    /// Represents the header structure of an FLV video tag.
+    /// </summary>
+    public record struct FlvTagHeader(FlvTagType TagType, uint DataSize, uint Timestamp)
     {
         public const int Size = 11;
 
@@ -14,12 +27,5 @@ namespace LiveStreamingServerNet.Flv.Internal
             dataBuffer.Write((byte)(Timestamp >> 24));
             dataBuffer.WriteUInt24BigEndian(0);
         }
-    }
-
-    internal enum FlvTagType : byte
-    {
-        Audio = 8,
-        Video = 9,
-        ScriptData = 18
     }
 }
